feat: added to appointment token card
